fix(model): avoid using asynchronous matrix de-normalization in `execute_or_add_commands`
In a previous version, de-normalization was performed asynchronously. However, this cause problems with concurrent file access, especially when de-normalizing a matrix (which can take time).
To avoid this concurrency problem, it would be necessary to implement a locking system for the entire study using a file lock (since multiple processes, not only multiple threads, could access the same content simultaneously).
Currently, we use a synchronous call to address the concurrency problem within the current process (not across multiple processes)...