feat(ui-modelization): add dynamic area selection on Areas tab click (#1835)
## Dynamic Handling of Area Selection in `Modelization` Component

### Description:

This PR introduces a significant change in the way areas are handled
within the `Modelization`.
The current system has a default area selection mechanism when
navigating to the map view. However, this approach has led to a couple
of issues:

- Inability to Access the Areas View without an Active Selection:
Previously, if a user navigated to the map view and deselected the
currently active area, they found themselves unable to navigate back to
the areas view unless they selected an area again. This behavior was
unintuitive and potentially frustrating for users who might want to
browse different areas or had no specific area to focus on initially.

- Errors When Creating New Layers with a Preselected Area:
The existing system also posed a problem when creating new layers while
an area was selected. If the selected area was not present in the newly
created layer, the application attempted to retrieve UI information for
the non-existent area in the context of that layer, leading to errors.

### Solution:

To address these issues, we have implemented a dynamic area handling
mechanism. Now, the area is determined dynamically when the "Areas" tab
is clicked. The key changes include:

- No Default Area on Initial Render:
No area is preselected. 

- Dynamic Area Selection on Tab Click:
The `"Areas"` tab click now triggers a dynamic process to determine and
set the current area. If no area is currently selected (i.e., `areaId`
is empty), the first area from the list of available areas is
automatically selected and dispatched to the Redux store. This selection
is then reflected in the navigation, taking the user to the view for the
selected area.