fix(ui-matrix): calculate the prepend index according to the existence of a time column (#1856)
This pull request addresses the column number offset issue in tables
without a time column on the left.

Changes in `EditableMatrix`:
- Modified the `prependIndex` variable to use an integer (0 or 1)
instead of a boolean to indicate the offset between the column number
being modified and the column index in the matrix.

Change in `utils.ts`:
- Added this parameter to the `cellChangesToMatrixEdits` function, which
converts `CellChange` objects to `MatrixEditDTO` objects, to account for
the column offset.

This fix resolves the problem for the "Overall Monthly Hydro" and "Daily
power" tables, which are not time series. No regressions have been
observed in the other time series tables.