feat(permission-db): check user permission through the db query (#1931)
Context: 
Additionally to ANT-1107 related tags db management, another problem of
permission filtering is blocking the pagination from working properly.

Issue: 
The permission to read studies in the search engine function is
performed after the db query. This is creating a problem for our
pagination process.

Solution:
Pass on the permission status directly in queries to the db.