fix(upgrade): raise an HTTP 417 exception when an upgrade has unmet requirements (#2047)
Context: 
This PR addresses two potential issues that can occur when upgrading
studies:
- Upgrading a variant study could potentially corrupt it.
- Upgrading a parent study that has variant children could potentially
corrupt those children.

Issue: 
At present, our system does not support the upgrade of commands, which
complicates the resolution of these issues. Until we develop a solution
for upgrading commands, this PR implements measures to prevent the two
situations mentioned above.

Solution: 
The modifications in this PR will trigger a 417 HTTP ERROR if an attempt
is made to upgrade a study that either has children or is a variant.
This is a temporary measure until we implement a solution for upgrading
commands.