ci(test): parallel pytest execution (#2133)
Enable the use of pytest-xdist to run tests in parallel, in CI
or developer env.

Following issues with parallel execution of tests are solved :

- some checks depended on the consistency at 2 moments in time of the
result of disk_usage, which was anyway not a good idea (dependent on
what happens elsewhere on the machine). It's now replaced with mocking
- some parameterized tests had random sort order, which is not allowed
by pytest-xdist (all workers must have exactly the same tests in the
same order)
- the command_factory test checked something in its teardown, assuming
all tests were executed in the same process, which is not the case
anymore. The same check is done as a separate unit tests.

Possible remaining issue which could cause some timeouts:
some services hold FileLock which have the same identifier
accross multiple tests.

Signed-off-by: Sylvain Leclerc <sylvain.leclerc@rte-france.com>