fix(desktop): fix urls for standalone version
The way we build the standalone version is completely
reviewed, because it does not work anymore with recent
versions of starlette.

The existing code was not correctly using root_path.
Actually, adding a prefix to the API requires to explicitly
add it to routes. It's a new configuration property "api_prefix"
(using fastapi sub-application could be an alternative solution).

Signed-off-by: Sylvain Leclerc <sylvain.leclerc@rte-france.com>