feat(watcher): filter out upgrade and TS generation `.tmp` folders (#2189)
Issue:
Temporary folders created during a study update are scanned by the
watcher, which imports studies into the database that we do not want to
import.

Solution:
To overcome this problem, we need to add a "hard coded" filter in the
watcher in addition to the "filter_out" filters that come from the
configuration, and other folders ignored by the watcher (AWS_NO_SCAN
etc). The convention of the upgrade is to create folders suffixed
".tmp": we can base ourselves on this convention. The best is probably
to create a constant in the upgrade module, in order to understand where
this filtering comes from.