feat(ui-studies): add on click fetch and display list of non studies folder (#2224)
There's a scan process that run on the server to update continuously the studies in the database.

However this process can take a long time, and the user shouldn't wait for hours before he can see a study he knows is already uploaded.

The scan button now by default run a non recursive scan, that way it takes less time. 
 
When the user clicks on a folder we fetch its subfolders using the explorer API. This enable the user to navigate to subfolders that aren't discovered by the scan yet, so subfolders that wouldn't be visible before hours if we rely only on the scan. 

By combining these two features, the user won't need to wait the full scan to complete, instead , he'll walk into the tree and run a fast scan process only subfolder he needs. 

This commit is mostly a front commit but also make small adjustment on the back.