refactor(commands): remove apply config (#2449)
apply_config was not tested in isolation and does not work
correctly at least for update_config.

It's actually used only to get the synthesis of a variant
study without generating a snapshot, but this is not a real
world use case. Actually, it generally made the /synthesis
endpoint very slow for variants, because we did not benefit
from the synthesis in cache.