feat(study-tree,ui-study-tree): preview studies not scanned yet (#2524)
The purpose of this feature is to show study folders in the study tree
with a badge next to the folder nem. This will help the user to identify
which folder contain a study that wasn't scanned yet so he can run a
scan on that folder.


![image](https://github.com/user-attachments/assets/d21b7903-49e1-43c0-b2ff-8825422f049c)

This feature is both on web and desktop but is essential to run the
desktop mode, as on dekstop there's no scan running in the background
the user needs to know what folder contains a study.

### Change summary 
**Backend** Updated the list_dir endpoint so it returns a FolderDTO
which can now be either a study folder or a non study folder, there's a
flag in the DTO for that.
**Frontend**: Show a badge next to study folder, updated the
TreeItemEnhanced to do so.

---------

Co-authored-by: Anis SMAIL <anis.smail_externe@rte-france.com>
Co-authored-by: Samir Kamal <1954121+skamril@users.noreply.github.com>