fix(desktop): clean deleted studies desktop mode (#2532)
Scanning is disabled in desktop mode, but we still need to remove
studies from the database if they no longer exist on disk.

We shouldn't rely on a full scan for this, as scanning all folders would
be too resource-intensive.

Instead, we'll iterate over the studies that have already been scanned
and check whether each one still exists on disk.

---------

Co-authored-by: Anis SMAIL <anis.smail_externe@rte-france.com>