fix(studies): make all managed studies folder end with id (#2600)
Fixes ANT-3640

Sometimes, a study's folder does not end with its own ID (as seen in the
top entry of the image below), even though this should always be the
case.

This inconsistency was observed in both staging and production
environments.

It is likely the result of a bug that was live for a few days several
weeks ago. The query below fixes the issue by appending the study's ID
to its folder path if it’s not already present.

One consequence of this inconsistency is that sometime, when the studies
are processed by the front end to build the study tree, depending on the
order on which they are processed some folders are marked a study
folders, they are then ignored by the front and they're not shown in the
study tree, for the user it seems that some folders are randomly missing
from the study tree. As you can see in the screenshot bellow the uae
folder isn't in the tree although it should.

---------

Co-authored-by: Anis SMAIL <anis.smail_externe@rte-france.com>