feat(matrix): introduce usage providers to get matrix usage info (#2583)
The MatrixUsageProvider abstraction is added so that clients of the
matrix service can register themselves to the service, and 
provide it with information about the matrices they refer to.

This allows to refactor the matrix GC to not rely on deep knowledge
of each client.
This will also allow to add further features which will collect stats
about used matrices.