fix(raw): fix getting link output data (#2660)
This is a quick-fix to revert a regression on getting some raw data.
Not great because it will cause some searches to be performed twice,
but at least it works.

The issue is that the previous code was relying on filesystem paths to
get
the relative "URL" of a node. This does not work, because paths and URLs
don't map 1 to 1. In particular, a link folder such as "de - fr" is
mapped to
a 2-levels URL `["de", "fr" ]`.

If we want to have a more robust solution we'll need to add a URL member
to
each node and update it correctly. Not sure it's worth it.

---------

Signed-off-by: Sylvain Leclerc <sylvain.leclerc@rte-france.com>