feat(layer): storage isolation (#2677)
Moves layer related methods to a new layer_manager. This manager
implements the DAO pattern.

Adds 3 commands :
- CreateLayer
- RemoveLayer
- UpdateLayerName

There is still something to do regarding layers, it is the updates of
all the areas associated with a given layer. However, this will be done
when the Area has been refactored with the new models.