fix(output)!: allow aggregation with columns mismatch (#2718)
**Breaking change**
Drops the support of HDF5 for output aggregation and matrix export

**Feature**
Supports parquet format for output aggregation and matrix export

**Fix**
- https://github.com/AntaresSimulatorTeam/antares_craft/issues/264
- If 2 areas/links did not share the same columns and we specifically
asked for a column only present in the second one, we used to stop and
return an empty response. Now we iterate over all files to find the
specified column

**Performance**:
- CSV format before: 18.8 seconds 
- CSV format now : 21.5s
- Parquet format with different columns: 14.8s
- Parquet format with same columns: 12.2s
- **Take-away**: 
     - We should use the parquet format.
- If possible we should request data for one area at a time or specify
which columns interest us. This way we'll take even greater advantage of
performance improvements