refactor(districts)!: introduce dao (#2686)
breaking change:

1. PUT /studies/{uuid}/districts/{district_id} doesn't accept extra
parameters anymore, request will be rejected if it contains other fields
than those defined in DistrictUpdate.
2. UpdateDistrict and CreateDistrict commands now accept a parameter
argument., which is an object that wraps the parameters for the command.
We can still pass parameters directly but we need to also pass an extra
parameter "version" set to 1.
3. The parameter filter_items was renamed to areas
4. the parameter base_filter was renamed to apply_filter
5. the fields set in FileStudyTreeConfigDTO and FileStudyTreeConfig were
renamed to districts


Other changes : 

1.  District object was introduced
2. DistrictSet was renamed DistrictFileData, also this class is not
referenced outside of the file where it's defined. FileStudyTreeConfig
now relies on District instead of DistrictDTO.

Bug fixes : 

1. updating a district attribute no longer erase areas
2. updating a district attribute no longer overwrite the add-all apply
filter
3. we handle studies with an "add-all" apply filter and a "+" attribute.
As well as studies wieh "remove-all" and "-".
4. caption are now optional. If a study has no caption on disk, loading
the district won't fail anymore.