feat(user): create a blobstore for variant user resources (#2777)
[ANT-3907]

**NOTE**: With the current behavior for variant and raw studies we're
writing data twice, once in the study and once in the blob store.
For variants, we have to as the snapshot will eventually be cleaned and
so we have to keep the data in the blob store.
For raw studies, we don't have to. The files will be removed from the
blob store the next time the garbage collector runs.

Also removes a useless loop inside the matrix garbage collector