feat(directory): implement directory management backend (#2774)
## New endpoints

- GET    /v1/directories                 # List all directories
- POST   /v1/directories               # Create a directory
- PATCH    /v1/directories/{id}         # update name or parent
- DELETE /v1/directories/{id}      # Delete directory if no study inside

 ## Database migration
- Create a new table `directory`
- Add a `directory_id` inside the study table
- All studies are parsed: every study that is in a specific folder will
trigger a creation of corresponding directories and will be assigned a
`directory_id`

 ## Rollback plan if needed
Just have to delete the new `directory` table and the new field in the
`study` table

## Transition from folder to directories

folder paths are built from the directory while the front-end dont use
the new directories endpoint. At some point, we want to remove this. I
have wrote a removal guide that will be needed when the front-end is
done (I have it locally).