fix(tasks): remove irrelevant status updates (#2802)
This PR removes 2 irrelevant task status updates:

- **timeouts:**
the timeout status of a task was set when a **waiting client** hit a
timeout,
which does not make sense since it's not the task itself which timed
out.
  In particular, the task execution continues and will anyway update the
  status at the end (either completed or failed)
- **cancels:**
a task can actually be cancelled only when it has not started. Once it
has
started, the execution cannot be interrupted. Therefore, it makes not
sense
to update the status in that case, since it will be overriden at the end
of the
  execution.

It also fixes the behaviour of task wait, which was not consistent:
in some cases we raised a `TimeoutError`, in others not.
This left the door open to weird behaviour, where one could consider a
task
finished whereas it was not.

---------

Signed-off-by: Sylvain Leclerc <sylvain.leclerc@rte-france.com>