fix(tasks): fix a memory leak (#2824)
The dictionary we use for storing futures is never ever cleaned up.
It seems it has never been an issue in production, but it still can
be non-negligible, in particular because it probably stores a reference
to the function executed, which generally carry some state with them.

Removing the future from the dict at the end of the computation
should not hurt, since the status is updated in database, therefore
future "waiters" will see on first request that it has indeed completed.

---------

Signed-off-by: Sylvain Leclerc <sylvain.leclerc@rte-france.com>