fix(study-ui): remove double fetch study call (#2869)
fix [ANT-4071](https://gopro-tickets.rte-france.com/browse/ANT-4071)

When opening a window, two requests to `/studies` and `/versions` were
triggered instead of one.

This happened because fetchStudies was dispatched in both` <Container
/>` and `<Studies />`.
This PR keeps the fetchStudies call only in <Container />.

`<UseAsyncAppSelectorCond />` was only used in `<Studies />`, so it’s
now unused and has been removed.

> If I switch between tabs, will the studies list still refresh?

The list of studies is already synced with the backend through a
WebSocket, so there's no need to fetch the studies list every time the
component loads.

---------

Co-authored-by: Anis SMAIL <anis.smail_externe@rte-france.com>