chore(logging): log server errors with request or task context (#2862)
Moves the logging of generic exceptions to the logging middleware.

Currently, they are logged by the exception handler, which is executed
**after** the logging
middleware, this without any request context. In particular, when we
look for logs of a
particular request, we don't get those error logs, which is very
inconvenient.

With that PR, we will correctly get the logs from those errors with the
trace_id of the request.

The PR also ensures tasks exceptions are logged in the task context, to
get the proper task_id.

---------

Signed-off-by: Sylvain Leclerc <sylvain.leclerc@rte-france.com>