fix(metrics): avoid arbitrary metrics creation (#2873)
The actual route is only resolved by fastapi after the first part of the
middleware, which makes it impossible to use for the http gauge metric, in
particular.

Also, all requests that don't get a proper route will now be labelled
with the others" endpoint.

This will avoid to have an explosion of the cardinality of timeseries in
prometheus.

Signed-off-by: Sylvain Leclerc <sylvain.leclerc@rte-france.com>