fix(tests): fix random test failures where study is not yet scanned (#2907)
Many integration tests rely on the scan of the STA-mini study, but we
don't
ensure that it is actually scanned before starting the tests.
This causes random failures of integration tests, in particular in
windows CI,
particularly since they have been made a little faster by avoiding
constant
matrix writing.

---------

Signed-off-by: Sylvain Leclerc <sylvain.leclerc@rte-france.com>