refactor(output): isolate the output storage (#2882)
The aim is to isolate the access to **output storage**, in order to be
able to propose a different implementation later. The separation of
concerns consists in:

1. isolating the **output service** from the **study services**:
- `OutputService` does not have any direct dependency to `StudyService`
    anymore (only access to necessary metadata)
- `IOutputStorage` implementation is now completely separate from
  `RawStudyService` and `VariantStudyService` (which provided the
   implementations, so far)
2. isolating the file **storage** related code from the **service**
    code, see `IOutputStorage` and its implementation `FileOutputStorage`.
   `OutputService` has no explicit reference anymore to the output file format.
   (the service code adds some app level logic like task management, events...)

In the future, the output service should be able to use another implementation
of `IOutputStorage`, which itself should be quite straight forward to implement.

The PR also separates the remote execution on workers from the task
service.
There was no real reason to mix those 2 features, and here it would have
forced us to introduce some task management at the storage level (for remote
execution of unarchiving). We now separate the remote execution from the task
handling.

---------

Signed-off-by: Sylvain Leclerc <sylvain.leclerc@rte-france.com>