refactor(matrix): perform less DB calls when normalizing/denormalizing (#2896)
Currently we perform 1 DB call for each matrix normalization or
denormalization.
For a BP study it means we're performing around 5000 DB calls !!!
With this PR we'll perform only 1 DB call for the whole study

We'll also see minor performance improvements inside the denormalization
process.
Before if we had 100 files that referenced the same matrix in the
matrix-store, we read its content 100 times.
Now we only read it once.